Update Anaconda Cluster -> Anaconda Adam for Repo channel mirroring
Replace Anaconda Cluster with Anaconda Adam channel for the Repo mirroring process. The `anaconda-adam` channel should be added to the tarballs for the airgapped installation instructions to be valid.

Note that mirroring the Anaconda Adam channel is only useful for core Adam functionality (i.e., remote conda and Salt) and does not imply that Adam can install *all* Anaconda platform components in an airgapped installation. More work needs to happen for that scenario.